feat: add logic for keyboard typing
