+++ index.css
@@ -4,14 +4,12 @@ html {
 }
 
 body {
-  font-family: "Georgia", sans-serif;
+  font-family: "Ysabeau", sans-serif;
   font-style: normal;
   font-weight: 400;
-  font-size: 1.5rem;
-  color: var(--color-dark-l);
-  /* letter-spacing: 0.04em; */
+  font-size: 1.6rem;
+  letter-spacing: 0.04em;
   margin: 0;
-  /* color-light-xl */
 
   background-color: #ffffff;
 }
@@ -47,6 +45,10 @@ img {
   height: auto;
 }
 
+.title {
+  text-align: center;
+}
+
 .container {
   margin: 0 auto;
   padding: 0 15px;
@@ -64,7 +66,7 @@ img {
   margin: 0 auto;
   padding-top: 20px;
   padding-bottom: 20px;
-  color: #f1f1f1;
+  color: #fff;
   background-color: #edc7c8;
 }
 
@@ -113,5 +115,17 @@ img {
 }
 
 .space {
-  width: 322px;
+  width: 324px;
+}
+
+.screen {
+  padding: 30px;
+}
+
+.screen__area {
+  display: block;
+  margin: 0 auto;
+  width: 800px;
+
+  resize: none;
 }
