+++ index.html
@@ -5,28 +5,20 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Ysabeau:wght@600&display=swap"
+      rel="stylesheet"
+    />
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/modern-normalize/1.1.0/modern-normalize.min.css"
     />
     <link rel="stylesheet" href="index.css" />
   </head>
-  <body>
-    <header class="container">
-      <h1>RSS virtual keyboard</h1>
-    </header>
-    <main class="container">
-      <section class="screen">
-        <div id="area" class="screen__area"></div>
-      </section>
-      <section class="keyboard">
-        <ul id="digit-line" class="keyboard__list"></ul>
-        <ul id="first-line" class="keyboard__list"></ul>
-        <ul id="second-line" class="keyboard__list"></ul>
-        <ul id="third-line" class="keyboard__list"></ul>
-        <ul id="fourth-line" class="keyboard__list special"></ul>
-      </section>
-    </main>
-    <script defer src="index.js" type="module"></script>
+  <body id="body">
+    <script defer src="/js/init-render.js" type="module"></script>
+    <script defer src="/js/index.js" type="module"></script>
   </body>
 </html>
